Implement the random hit function for the proxy
The idea is that the proxy will ensure the creation of the tunnel from
the data node to the master node. This tunnel will return a
local_bind_port which links the proxy to the data node.

This way when we target the localhost on the proxy, the request will be
forwarded to the data node which is bound to the master node.

Signed-off-by: Mark Bekhet <markonsy72@gmail.com>